# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "An API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3003",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/maker/id/{id}": {
      "get": {
        "tags": [
          "Makers"
        ],
        "description": "Maker endpoints",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Maker's ID"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Maker"
            },
            "description": "A Maker!"
          }
        }
      }
    },
    "/api/maker/name/{name}": {
      "get": {
        "tags": [
          "Makers"
        ],
        "description": "Maker endpoints",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Maker's short-name"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Maker"
            },
            "description": "A Maker!"
          }
        }
      }
    },
...
    "/api/vendor/id/{id}": {
      "get": {
        "tags": [
          "Vendors"
        ],
        "description": "Vendor endpoints",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Vendor's ID"
          }
        ],
        "responses": {}
      }
    },
    "/api/vendor/name/{name}": {
      "get": {
        "tags": [
          "Vendors"
        ],
        "description": "Vendor endpoints",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Vendor's short-name"
          }
        ],
        "responses": {}
      }
    },
...
    "/api/purchase/{id}": {
      "get": {
        "tags": [
          "Purchases"
        ],
        "description": "Purchase endpoints",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Purchase ID"
          }
        ],
        "responses": {}
      }
    },
    "/api/orderset/{id}": {
      "get": {
        "tags": [
          "Purchases"
        ],
        "description": "Purchase endpoints",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "OrderSet ID"
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {
    "Maker": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "acme-maker"
        },
        "display_name": {
          "type": "string",
          "example": "ACME Maker"
        },
        "instagram": {
          "type": "string",
          "example": "acmemaker"
        }
      }
    },
...
  }
}